+++ packages/deepseek-llm/src/index.ts
@@ -1,0 +1,2 @@
+export type { DeepSeekConfig } from './deepseek-provider.js';
+export { DeepSeekProvider } from './deepseek-provider.js';
