+++ packages/llm-agent-server/src/index.ts
@@ -9,6 +9,10 @@ export {
   type AnthropicConfig,
   AnthropicProvider,
 } from '@mcp-abap-adt/anthropic-llm';
+export {
+  type DeepSeekConfig,
+  DeepSeekProvider,
+} from '@mcp-abap-adt/deepseek-llm';
 export { BaseLLMProvider, type LLMProvider } from '@mcp-abap-adt/llm-agent';
 export { type OpenAIConfig, OpenAIProvider } from '@mcp-abap-adt/openai-llm';
 // Re-export for backward compatibility
@@ -42,10 +46,6 @@ export {
   SapCoreAIAgent,
   type SapCoreAIAgentConfig,
 } from './agents/sap-core-ai-agent.js';
-export {
-  type DeepSeekConfig,
-  DeepSeekProvider,
-} from './llm-providers/deepseek.js';
 // MCP Client
 export {
   type MCPClientConfig,
